+++ src/main/java/terramine/common/utility/CorruptionHelper.java
@@ -70,22 +70,20 @@ public class CorruptionHelper extends SpreadingSnowyDirtBlock  {
             BlockState grass = ModBlocks.CORRUPTED_GRASS.defaultBlockState();
             BlockState snow_layer = ModBlocks.CORRUPTED_SNOW_LAYER.defaultBlockState();
 
-            for (int i = 0; i < 4; ++i) { // corrupted grass spread to grass
-                BlockPos blockPos2 = blockPos.offset(random.nextInt(3) - 1, random.nextInt(5) - 3, random.nextInt(3) - 1);
-                if (!serverLevel.getBlockState(blockPos2).is(Blocks.GRASS_BLOCK) || canNotPropagate(grass, serverLevel, blockPos2)) continue;
-                serverLevel.setBlockAndUpdate(blockPos2, grass.setValue(SNOWY, (serverLevel.getBlockState(blockPos2.above()).is(Blocks.SNOW) || serverLevel.getBlockState(blockPos2.above()).is(ModBlocks.CORRUPTED_SNOW_LAYER))));
-                spreadBiome(serverLevel, blockPos2, false);
-            }
-            for (int i = 0; i < 4; ++i) { // normal grass spread with corrupted grass
-                BlockPos blockPos2 = blockPos.offset(random.nextInt(3) - 1, random.nextInt(5) - 3, random.nextInt(3) - 1);
-                if (!serverLevel.getBlockState(blockPos2).is(Blocks.DIRT) || canNotPropagate(grass, serverLevel, blockPos2)) continue;
-                serverLevel.setBlockAndUpdate(blockPos2, grass.setValue(SNOWY, (serverLevel.getBlockState(blockPos2.above()).is(Blocks.SNOW) || serverLevel.getBlockState(blockPos2.above()).is(ModBlocks.CORRUPTED_SNOW_LAYER))));
-                spreadBiome(serverLevel, blockPos2, false);
+            for (int i = 0; i < 4; ++i) { // corrupted grass spread to grass and dirt
+                if (random.nextInt(TerraMine.CONFIG.general.evilSpreadRarity + 1) == 1) {
+                    BlockPos blockPos2 = blockPos.offset(random.nextInt(3) - 1, random.nextInt(3) - 1, random.nextInt(3) - 1);
+                    if ((!serverLevel.getBlockState(blockPos2).is(Blocks.GRASS_BLOCK) && !serverLevel.getBlockState(blockPos2).is(Blocks.DIRT)) || canNotPropagate(grass, serverLevel, blockPos2)) continue;
+                    serverLevel.setBlockAndUpdate(blockPos2, grass.setValue(SNOWY, (serverLevel.getBlockState(blockPos2.above()).is(Blocks.SNOW) || serverLevel.getBlockState(blockPos2.above()).is(ModBlocks.CORRUPTED_SNOW_LAYER))));
+                    spreadBiome(serverLevel, blockPos2, false);
+                }
             }
             for (int i = 0; i < 4; ++i) { // spread layered snow
-                BlockPos blockPos2 = blockPos.offset(random.nextInt(3) - 1, random.nextInt(5) - 3, random.nextInt(3) - 1);
-                if (!serverLevel.getBlockState(blockPos2).is(Blocks.SNOW)) continue;
-                serverLevel.setBlockAndUpdate(blockPos2, snow_layer.setValue(CorruptedSnowLayer.LAYERS, serverLevel.getBlockState(blockPos2).getValue(SnowLayerBlock.LAYERS)));
+                if (random.nextInt(TerraMine.CONFIG.general.evilSpreadRarity + 1) == 1) {
+                    BlockPos blockPos2 = blockPos.offset(random.nextInt(3) - 1, random.nextInt(3) - 1, random.nextInt(3) - 1);
+                    if (!serverLevel.getBlockState(blockPos2).is(Blocks.SNOW)) continue;
+                    serverLevel.setBlockAndUpdate(blockPos2, snow_layer.setValue(CorruptedSnowLayer.LAYERS, serverLevel.getBlockState(blockPos2).getValue(SnowLayerBlock.LAYERS)));
+                }
             }
 
             spreadBlock(ModBlocks.CORRUPTED_GRAVEL, Blocks.GRAVEL, serverLevel, blockPos, random);
@@ -126,7 +124,7 @@ public class CorruptionHelper extends SpreadingSnowyDirtBlock  {
         for (int i = 0; i < 4; ++i) {
             if (random.nextInt(TerraMine.CONFIG.general.evilSpreadRarity + 1) == 1) {
                 BlockState block = toSpread.defaultBlockState();
-                BlockPos blockPos2 = blockPos.offset(random.nextInt(3) - 1, random.nextInt(5) - 3, random.nextInt(3) - 1);
+                BlockPos blockPos2 = blockPos.offset(random.nextInt(3) - 1, random.nextInt(3) - 1, random.nextInt(3) - 1);
                 if (!serverLevel.getBlockState(blockPos2).is(spreadTo)) continue;
                 serverLevel.setBlockAndUpdate(blockPos2, block.setValue(SNOWY, (serverLevel.getBlockState(blockPos2.above()).is(Blocks.SNOW) || serverLevel.getBlockState(blockPos2.above()).is(ModBlocks.CORRUPTED_SNOW_LAYER))));
                 spreadBiome(serverLevel, blockPos2, false);